backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-multipart==0.0.9
openai==1.51.0
httpx<0.28,>=0.27
python-dotenv>=1.0.1
